requirements.txt
# FastAPI: The web framework for building APIs
fastapi

# Uvicorn: ASGI server for running FastAPI applications
uvicorn

# SQLAlchemy: ORM for database interactions
sqlalchemy

# Pydantic: Data validation and settings management
pydantic

# Passlib with bcrypt: Secure password hashing
passlib[bcrypt]

# python-jose with cryptography: JWT creation and verification
python-jose[cryptography]